refactored wx features
